refactor(lib): extract AttendanceStateManager and CSV exporter (phase 2)
- Move AttendanceStateManager to lib.state; lazy-import where used
- Introduce lib.csv_exporter with header/status/rows helpers and save function
- Delegate export_csv to csv_exporter; behavior unchanged (tests pass)
